Heavy structural changes
+ Used Strategy design pattern to dynamically set the CardSequence type
through an implementation of the CardSequenceType interface that will be
passed to the constructor of the CardSequence class
+ Used Builder design pattern to properly instantiate a CardSequence
object. Only one builder class is needed since there is only one
CardSequence implementation thus only one constructor, which was the
whole reason why there could not be a single generic builder for
different GenericCardSequence (abstract class) subclasses. (Actually
there were ways, but they were to messy)
+ Documentation for CardSequence, CardSequenceType,
CardSequenceListener, CardSequenceCaller and CardSequenceBuilder
* Other necessary modifications
* Fixed tests to work with modifications